don't remove files from history for which there are still favorites

+++ src/FileHistory.h
@@ -67,10 +67,12 @@ public:
     FileHistory() { }
     ~FileHistory() { }
 
-    void Clear() {
+    void Clear(bool keepFavorites) {
         if (!states)
             return;
         for (size_t i = 0; i < states->Count(); i++) {
+            if (keepFavorites && states->At(i)->favorites->Count() > 0)
+                continue;
             DeleteDisplayState(states->At(i));
         }
         states->Reset();
